+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 4.77.0 - (September 20, 2022)
+
+* Changed
+  * Fixed Keyboard navigation when moving over the start day of DST.
+
 ## 4.76.0 - (July 27, 2022)
 
 * Changed