+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## Unreleased
 
+## 1.10.0 - (September 20, 2022)
+
 * Changed
   * Updated doc and added test for terra-menu
   * Updated doc and added test for terra-tabs
+  * Updated DST example for `terra-date-time-picker`.
 
 ## 1.9.0 - (July 27, 2022)